+++ packages/article/src/Domain/Model/ArticleDimensionContent.php
@@ -23,7 +23,6 @@ use Sulu\Content\Domain\Model\SeoTrait;
 use Sulu\Content\Domain\Model\ShadowTrait;
 use Sulu\Content\Domain\Model\TaxonomyTrait;
 use Sulu\Content\Domain\Model\TemplateTrait;
-//use Sulu\Content\Domain\Model\WebspaceTrait;
 use Sulu\Content\Domain\Model\WebspaceTrait;
 use Sulu\Content\Domain\Model\WorkflowTrait;
