+++ packages/content/config/resolvers.xml
@@ -61,11 +61,12 @@
             <tag name="sulu_content.content_resolver" type="settings"/>
         </service>
 
-        <!-- Excerpt Resolver -->
-        <service id="sulu_content.excerpt_resolver"
-                 class="Sulu\Content\Application\ContentResolver\Resolver\ExcerptResolver">
+        <!-- Excerpt and Taxonomy Resolver -->
+        <service id="sulu_content.excerpt_taxonomy_resolver"
+                 class="Sulu\Content\Application\ContentResolver\Resolver\ExcerptTaxonomyResolver">
             <argument type="service" id="sulu_admin.form_metadata_provider"/>
             <argument type="service" id="sulu_content.metadata_resolver"/>
+            <argument>%sulu_content.content_excerpt_forms%</argument>
 
             <tag name="sulu_content.content_resolver" type="excerpt"/>
         </service>
