+++ packages/content/src/Application/ContentMerger/Merger/ExcerptMerger.php
@@ -14,7 +14,6 @@ declare(strict_types=1);
 namespace Sulu\Content\Application\ContentMerger\Merger;
 
 use Sulu\Content\Domain\Model\ExcerptInterface;
-use Sulu\Content\Domain\Model\TaxonomyInterface;
 
 /**
  * @internal This class should not be instantiated by a project.
@@ -50,32 +49,6 @@ final class ExcerptMerger implements MergerInterface
 
         if ($excerptImage = $sourceObject->getExcerptImage()) {
             $targetObject->setExcerptImage($excerptImage);
-        }
-
-        if (!$targetObject instanceof TaxonomyInterface || !$sourceObject instanceof TaxonomyInterface) {
-            return;
-        }
-
-        if ($excerptSegment = $sourceObject->getExcerptSegment()) {
-            $targetObject->setExcerptSegment($excerptSegment);
-        }
-
-        if ($excerptTags = $sourceObject->getExcerptTags()) {
-            if (\count($excerptTags) > 0) { // @phpstan-ignore-line false positive for phpstan thinks it is a non-empty-array
-                $targetObject->setExcerptTags($excerptTags);
-            }
-        }
-
-        if ($excerptCategories = $sourceObject->getExcerptCategories()) {
-            if (\count($excerptCategories) > 0) { // @phpstan-ignore-line false positive for phpstan thinks it is a non-empty-array
-                $targetObject->setExcerptCategories($excerptCategories);
-            }
-        }
-
-        if ($excerptAudienceTargetGroups = $sourceObject->getExcerptAudienceTargetGroups()) {
-            if (\count($excerptAudienceTargetGroups) > 0) { // @phpstan-ignore-line false positive for phpstan thinks it is a non-empty-array
-                $targetObject->setExcerptAudienceTargetGroups($excerptAudienceTargetGroups);
-            }
         }
     }
 }
