+++ packages/content/src/Infrastructure/Doctrine/MetadataLoader.php
@@ -29,6 +29,7 @@ use Sulu\Content\Domain\Model\LinkInterface;
 use Sulu\Content\Domain\Model\RoutableInterface;
 use Sulu\Content\Domain\Model\SeoInterface;
 use Sulu\Content\Domain\Model\ShadowInterface;
+use Sulu\Content\Domain\Model\TaxonomyInterface;
 use Sulu\Content\Domain\Model\TemplateInterface;
 use Sulu\Content\Domain\Model\WebspaceInterface;
 use Sulu\Content\Domain\Model\WorkflowInterface;
@@ -93,7 +94,6 @@ final class MetadataLoader
             $this->addField($metadata, 'excerptTitle');
             $this->addField($metadata, 'excerptMore', 'string', ['length' => 63]);
             $this->addField($metadata, 'excerptDescription', 'text');
-            $this->addField($metadata, 'excerptSegment');
             $this->addField($metadata, 'excerptImageId', 'integer', [
                 'columnName' => 'excerptImageId',
                 '_custom' => [
@@ -115,7 +115,10 @@ final class MetadataLoader
                     ],
                 ],
             ]);
+        }
 
+        if ($reflection->implementsInterface(TaxonomyInterface::class)) {
+            $this->addField($metadata, 'excerptSegment');
             $this->addManyToMany($event, $metadata, 'excerptTags', TagInterface::class, 'tag_id');
             $this->addManyToMany($event, $metadata, 'excerptCategories', CategoryInterface::class, 'category_id');
             if ($this->bundles['SuluAudienceTargetingBundle'] ?? false) {
