+++ packages/content/tests/Unit/Content/Application/ContentNormalizer/ContentNormalizerTest.php
@@ -19,8 +19,8 @@ use Sulu\Bundle\TagBundle\Tag\TagInterface;
 use Sulu\Content\Application\ContentNormalizer\ContentNormalizer;
 use Sulu\Content\Application\ContentNormalizer\ContentNormalizerInterface;
 use Sulu\Content\Application\ContentNormalizer\Normalizer\DimensionContentNormalizer;
-use Sulu\Content\Application\ContentNormalizer\Normalizer\ExcerptNormalizer;
 use Sulu\Content\Application\ContentNormalizer\Normalizer\RoutableNormalizer;
+use Sulu\Content\Application\ContentNormalizer\Normalizer\TaxonomyNormalizer;
 use Sulu\Content\Application\ContentNormalizer\Normalizer\TemplateNormalizer;
 use Sulu\Content\Application\ContentNormalizer\Normalizer\WorkflowNormalizer;
 use Sulu\Content\Domain\Model\ContentRichEntityInterface;
@@ -47,7 +47,7 @@ class ContentNormalizerTest extends TestCase
     {
         return new ContentNormalizer([
             new DimensionContentNormalizer(),
-            new ExcerptNormalizer(),
+            new TaxonomyNormalizer(),
             new TemplateNormalizer(),
             new WorkflowNormalizer(),
             new RoutableNormalizer(),
