+++ packages/snippet/tests/Functional/Integration/responses/snippet_get_after_restore.json
@@ -8,15 +8,10 @@
     ],
     "description": null,
     "excerptCategories": [],
-    "excerptDescription": "Excerpt Description",
-    "excerptIcon": null,
-    "excerptImage": null,
-    "excerptMore": "Excerpt More",
     "excerptTags": [
         "Tag 1",
         "Tag 2"
     ],
-    "excerptTitle": "Excerpt Title",
     "excerptAudienceTargetGroups": [],
     "excerptSegment": null,
     "ghostLocale": "en",
