+++ packages/snippet/tests/Functional/Integration/responses/snippet_get_ghost_locale.json
@@ -5,12 +5,7 @@
     "created": "@datetime@",
     "creator" : "@integer@",
     "excerptCategories": [],
-    "excerptDescription": null,
-    "excerptIcon": null,
-    "excerptImage": null,
-    "excerptMore": null,
     "excerptTags": [],
-    "excerptTitle":null,
     "excerptAudienceTargetGroups": [],
     "excerptSegment": null,
     "ghostLocale": "en",
