+++ packages/snippet/tests/Functional/Integration/responses/snippet_post_publish.json
@@ -9,16 +9,11 @@
     "description": null,
     "excerptAudienceTargetGroups": [],
     "excerptCategories": [],
-    "excerptDescription": "Excerpt Description",
-    "excerptIcon": null,
-    "excerptImage": null,
-    "excerptMore": "Excerpt More",
     "excerptSegment": null,
     "excerptTags": [
         "Tag 1",
         "Tag 2"
     ],
-    "excerptTitle": "Excerpt Title",
     "ghostLocale": "en",
     "id": "@string@",
     "image": null,
