+++ packages/snippet/tests/Functional/Integration/responses/snippet_post_restore.json
@@ -10,16 +10,11 @@
     "description": "<p>Test Snippet 2</p>",
     "excerptAudienceTargetGroups": [],
     "excerptCategories": [],
-    "excerptDescription": "Excerpt Description 2",
-    "excerptIcon": null,
-    "excerptImage": null,
-    "excerptMore": "Excerpt More 2",
     "excerptSegment": null,
     "excerptTags": [
         "Tag 3",
         "Tag 4"
     ],
-    "excerptTitle": "Excerpt Title 2",
     "ghostLocale": "en",
     "id": "@string@",
     "image": null,
