+++ packages/snippet/tests/Functional/Integration/responses/snippet_post_trigger_copy_locale.json
@@ -9,15 +9,10 @@
     "creator" : "@integer@",
     "description": null,
     "excerptCategories": [],
-    "excerptDescription": "Excerpt Description",
-    "excerptIcon": null,
-    "excerptImage": null,
-    "excerptMore": "Excerpt More",
     "excerptTags": [
         "Tag 1",
         "Tag 2"
     ],
-    "excerptTitle": "Excerpt Title",
     "excerptAudienceTargetGroups": [],
     "excerptSegment": null,
     "ghostLocale": "en",
