+++ packages/snippet/tests/Functional/Integration/responses/snippet_post_trigger_unpublish.json
@@ -8,15 +8,10 @@
     "creator" : "@integer@",
     "description": null,
     "excerptCategories": [],
-    "excerptDescription": "Excerpt Description",
-    "excerptIcon": null,
-    "excerptImage": null,
-    "excerptMore": "Excerpt More",
     "excerptTags": [
         "Tag 1",
         "Tag 2"
     ],
-    "excerptTitle": "Excerpt Title",
     "excerptAudienceTargetGroups": [],
     "excerptSegment": null,
     "ghostLocale": "en",
