+++ packages/snippet/tests/Functional/Integration/responses/snippet_put.json
@@ -9,15 +9,10 @@
     "creator" : "@integer@",
     "description": "<p>Test Snippet 2</p>",
     "excerptCategories": [],
-    "excerptDescription": "Excerpt Description 2",
-    "excerptIcon": null,
-    "excerptImage": null,
-    "excerptMore": "Excerpt More 2",
     "excerptTags": [
         "Tag 3",
         "Tag 4"
     ],
-    "excerptTitle": "Excerpt Title 2",
     "excerptAudienceTargetGroups": [],
     "excerptSegment": null,
     "ghostLocale": "en",
